backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
channels-redis==4.2.0
django==5.2
django-cors-headers==4.3.1
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
redis==5.2.1
whitenoise==6.8.0
openpyxl==3.1.5
transformers==4.51.3
torch==2.6.0
numpy==1.26.4
scikit-learn==1.6.1
pandas==2.2.3
openai==1.74.0

safetensors==0.5.3
tokenizers==0.21.1
tqdm==4.67.1
regex==2024.11.6
filelock==3.18.0
PyYAML==6.0.2
reportlab==4.1.0
django==5.2
huggingface_hub==0.30.2
python-dotenv==1.0.1
requests==2.32.3
together==1.3.1